requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-dotenv==1.0.0
google-cloud-aiplatform==1.26.1
google-cloud-storage==2.8.0
numpy==1.24.3
faiss-cpu==1.7.4
Jinja2==3.1.2